Trust_System/backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
sqlalchemy>=2.0,<3.0
asyncpg>=0.29,<1.0
alembic>=1.13,<2.0
pydantic-settings>=2.5,<3.0
python-jose[cryptography]>=3.3,<4.0
passlib[bcrypt]>=1.7,<2.0
email-validator>=2.0,<3.0
